requirements.txt
gym==0.26.0
numpy==1.25.2
pygame==2.1.0
PyYAML==6.0.1
pytest==8.0.0
tensorflow==2.15.0